Refactored the VSCode extension to allow it to run even if the language server fails
