+++ clients/oraide-vscode/src/extension.ts
@@ -20,10 +20,6 @@ export async function activate(context: vscode.ExtensionContext) {
         logger.appendLine("This is running for real!");
     }
 
-    // Checking if the fallback OpenRA folder is configured.
-    const config = vscode.workspace.getConfiguration();
-    const gameFolderPath = config.get<string>('oraide.game.path');
-
     // Just a sanity check for the workspace folder.
     if (vscode.workspace.workspaceFolders === undefined || vscode.workspace.workspaceFolders.length === 0) {
         vscode.window.showInformationMessage('Something went wrong. Extension ORAIDE will abort.');
@@ -32,34 +28,17 @@ export async function activate(context: vscode.ExtensionContext) {
 
     // Explicitly not supporting multi-root workspaces for the moment. May change later depending on requirements.
     // https://code.visualstudio.com/docs/editor/multi-root-workspaces
+    // TODO: What if the workspace changes during execution??
     const workspaceFolderPath = vscode.workspace.workspaceFolders[0].uri.fsPath;
 
-    // TODO: What if the workspace changes during execution??
-
-    // Locate language server binary.
-    let serverPath = await languageServer.getLanguageServerPath(context, config);
-    if (!serverPath) {
-        logger.appendLine("Extension ORAIDE failed to find or download its language server and will abort.");
-        logger.appendLine("If you are running in debug mode, configure the path to the language server in the extension settings!");
-        vscode.window.showInformationMessage("Extension ORAIDE failed to find or download its language server and will abort.");
-        return;
+    // Try to run the language server.
+    const config = vscode.workspace.getConfiguration();
+    let result = await languageServer.tryStart(context, config, workspaceFolderPath);
+    if (result.isSuccessful) {
+        client = result.client!;
     }
 
-    start(context, serverPath, workspaceFolderPath, gameFolderPath!);
-
-    // Detect if a file identified as YAML is actually MiniYAML and switch if so.
-    function switchLanguageIfMiniYAML(doc: vscode.TextDocument) {
-        if (doc.languageId == 'yaml') {
-            const text = doc.getText();
-            if (text.includes('\t')) {
-                vscode.languages.setTextDocumentLanguage(doc, 'miniyaml');
-            }
-        }
-    }
-    for (const editor of vscode.window.visibleTextEditors) {
-        switchLanguageIfMiniYAML(editor.document);
-    }
-    vscode.workspace.onDidOpenTextDocument(switchLanguageIfMiniYAML);
+    handleMiniYamlFiles();
 }
 
 // this method is called when your extension is deactivated
@@ -73,43 +52,20 @@ export function deactivate() {
     return client.stop();
 }
 
-function start(context: vscode.ExtensionContext, serverPath: string, workspaceFolderPath: string, defaultOpenRaPath: string) {
-    
-    logger.appendLine(`METHOD start`);
-    logger.appendLine(`  serverPath: ${serverPath}`);
-    logger.appendLine(`  workspaceFolderPath: ${workspaceFolderPath}`);
-    logger.appendLine(`  defaultOpenRaPath: ${defaultOpenRaPath}`);
-    
-    const serverOptions: vscodelc.ServerOptions = async () => languageServer.spawnServerProcess(serverPath, workspaceFolderPath, defaultOpenRaPath);
+function handleMiniYamlFiles() {
+    // Detect if a file identified as YAML is actually MiniYAML and switch if so.
+    function switchLanguageIfMiniYAML(doc: vscode.TextDocument) {
+        if (doc.languageId == 'yaml') {
+            const text = doc.getText();
+            if (text.includes('\t')) {
+                vscode.languages.setTextDocumentLanguage(doc, 'miniyaml');
+            }
+        }
+    }
 
-    const clientOptions: vscodelc.LanguageClientOptions = {
-        // Register the server for 'yaml' (.yaml) files. This uses the definition for 'miniyaml' found in package.json under 'contributes.languages'.
-        documentSelector: [
-            {
-                language: 'miniyaml',
-                scheme: 'file',
-            },
-            {
-                language: 'miniyaml',
-                scheme: 'untitled',
-            },
-        ],
-        diagnosticCollectionName: 'OpenRA IDE',
-        synchronize: {
-            // Notify the server about file changes to '.yaml files contained in the workspace
-            configurationSection: 'oraide',
-            fileEvents: vscode.workspace.createFileSystemWatcher('**/*.yaml')
-        },
-        outputChannelName: "OpenRA IDE server",
-    };
+    for (const editor of vscode.window.visibleTextEditors) {
+        switchLanguageIfMiniYAML(editor.document);
+    }
 
-    let client = new vscodelc.LanguageClient('oraide', 'OpenRA IDE', serverOptions, clientOptions, true);
-
-    let disposable = client.start();
-
-    logger.appendLine("Started client")
-
-    // Push the disposable to the context's subscriptions so that the
-    // client can be deactivated on extension deactivation
-    context.subscriptions.push(disposable);
+    vscode.workspace.onDidOpenTextDocument(switchLanguageIfMiniYAML);
 }
